Add files via upload
This a game which calls the player to guess for a secret number between 1-20. The initial score is 20 and for every wrong guess a value of 1 is subtracted from the score until it reaches zero and the player loses the game. The purpose is to acheive the highest score possible.   The game uses HTML, CSS and JAVASCRIPT in order to manipulate the DOM and enable the user to play the game.